introduce streaming table
