+++ src/storage/src/streaming_table/mem_table.rs
@@ -15,10 +15,6 @@ use std::collections::btree_map::Entry;
 use std::collections::BTreeMap;
 use std::ops::RangeBounds;
 
-use risingwave_common::array::Row;
-
-use crate::error::StorageResult;
-
 #[derive(Clone)]
 pub enum RowOp {
     Insert(Vec<u8>),
@@ -147,14 +143,5 @@ impl MemTable {
         R: RangeBounds<Vec<u8>> + 'a,
     {
         self.buffer.range(key_range)
-    }
-
-    /// Write to state store.
-    pub async fn batch_write_rows(
-        &mut self,
-        buffer: BTreeMap<Vec<u8>, RowOp>,
-        epoch: u64,
-    ) -> StorageResult<()> {
-        todo!()
     }
 }
